Move utility functions to common folder
